Refactor cut_plotter into cut_plotter and cut_algorithm
