+++ models/cut/cut_algorithm.py
@@ -1,0 +1,5 @@
+
+
+class CutAlgorithm(object):
+    def compute_cut(self, selected_workspace, cut_axis, integration_start, integration_end):
+        pass
